feat(wamuseum-client): :art: ui convention for features add tickets
