+++ apps/wamuseum-client/app/concert/[id]/page.client.tsx
@@ -1,5 +1,6 @@
 'use client'
 
+import { AddTicketsUI } from '@/features'
 import useCreateConcertPosterMutation from '@/hooks/useCreateConcertPosterMutation'
 import useUpdateConcertPosterMutation from '@/hooks/useUpdateConcertPosterMutation'
 import { presign, uploadToPresignedURL } from '@/utils/fetcher'
@@ -12,7 +13,6 @@ import { concertPosterQuery } from 'gql/queries'
 import { useRouter } from 'next/navigation'
 import { useCallback, useMemo, useState } from 'react'
 import useConcertQuery from '../../../hooks/useConcertQuery'
-import AddTicketsUI from './components/AddTicketsUI'
 import PosterUI from './components/PosterUI'
 import RegisteredArtist from './components/RegisteredArtist'
 import RegisteredTicketsUI from './components/RegisteredTicketsUI'
