+++ apps/wamuseum-client/features/create-concert/ui/add-tickets-ui/add-tickets-ui.styled.ts
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled'
+
+export const StyledAddTicketsUIContent = styled.h3`
+  font-size: 20px;
+  margin-bottom: 16px;
+  font-weight: 600;
+  display: flex;
+  align-items: center;
+`
